requirements.txt
openai==1.30.2
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
python-dotenv==1.0.1
opencv-python==4.10.0.84
textract 1.6.5
